:recycle: extract shared GitHub label parsing to commons.github
